requirements.txt
# Core dependencies
aiohttp==3.9.5
sqlalchemy>=2.0.0
APScheduler==3.10.4
colorama==0.4.6
dependency-injector==4.41.0
docker==7.1.0
Flask==3.0.1
Jinja2==3.1.2
numpy==1.26.4
ollama==0.6.1
pydantic==2.12.5
pydantic-settings==2.2.1
python-dotenv==1.0.0
PyYAML==6.0.1
pytz==2023.3
requests==2.32.5
rich==13.7.1
prompt_toolkit>=3.0.52
tiktoken==0.6.0
Werkzeug==3.0.1

# Async support
aiofiles==23.2.1
aiosqlite==0.19.0

# FastAPI + ASGI server (replaces Flask/Werkzeug)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# WebSocket support for terminal (keep for CLI compatibility)
flask-sock==0.7.0

# Type checking support
typing_extensions>=4.8.0

# System monitoring
psutil==5.9.8

# Document processing
pypdf==4.2.0
python-docx==1.0.1
python-pptx==0.6.23

# Auth (local JWT + bcrypt + rate limiting)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9
